+++ src/common/KmerAnalysis.h
@@ -10,7 +10,7 @@ int get_coverage(std::vector<ReadFileMetaData> &read_meta_data, int genome_size)
 
 int get_genome_size(std::vector<ReadFileMetaData> &read_meta_data, int coverage);
 
-std::vector<uint8_t> get_k_sizes(int max_genome_size);
+std::vector<unsigned int> get_k_sizes(int max_genome_size);
 
 uint64_t get_approximate_kmer_count(SequenceRecordIterator &read_iterator, int k);
